Move the test into right namespace.
