+++ osu.Game.Rulesets.Karaoke.Tests/Screens/Edit/Beatmap/Lyrics/Settings/TestSceneLyricEditorDescriptionTextFlowContainer.cs
@@ -13,7 +13,7 @@ using osu.Game.Rulesets.Karaoke.Screens.Edit.Components.Markdown;
 using osu.Game.Tests.Visual;
 using osuTK;
 
-namespace osu.Game.Rulesets.Karaoke.Tests.Screens.Edit.Beatmap.Lyrics.Components;
+namespace osu.Game.Rulesets.Karaoke.Tests.Screens.Edit.Beatmap.Lyrics.Settings;
 
 [TestFixture]
 public partial class TestSceneDescriptionTextFlowContainer : OsuTestScene
